refactored, commented, tested connectivity scripts
